move getPersistenceAdapter(tableName) function to util.js
