+++ 04/lambda/index.js
@@ -1,5 +1,5 @@
 const Alexa = require('ask-sdk-core');
-var persistenceAdapter = getPersistenceAdapter();
+const util = require('./util');
 
 const moment = require('moment-timezone'); // will help us do all the birthday math
 
@@ -10,26 +10,6 @@ const i18n = require('i18next');
 // The keys for each string will then be referenced in our code
 // e.g. handlerInput.t('WELCOME_MSG')
 const languageStrings = require('./localisation');
-
-function getPersistenceAdapter(tableName) {
-    // This function is an indirect way to detect if this is part of an Alexa-Hosted skill
-    function isAlexaHosted() {
-        return process.env.S3_PERSISTENCE_BUCKET;
-    }
-    if(isAlexaHosted()) {
-        const {S3PersistenceAdapter} = require('ask-sdk-s3-persistence-adapter');
-        return new S3PersistenceAdapter({
-            bucketName: process.env.S3_PERSISTENCE_BUCKET
-        });
-    } else {
-        // IMPORTANT: don't forget to give DynamoDB access to the role you're to run this lambda (IAM)
-        const {DynamoDbPersistenceAdapter} = require('ask-sdk-dynamodb-persistence-adapter');
-        return new DynamoDbPersistenceAdapter({
-            tableName: tableName,
-            createTable: true
-        });
-    }
-}
 
 const LaunchRequestHandler = {
     canHandle(handlerInput) {
@@ -292,5 +272,5 @@ exports.handler = Alexa.SkillBuilders.custom()
     .addResponseInterceptors(
         LoggingResponseInterceptor,
         SaveAttributesResponseInterceptor)
-    .withPersistenceAdapter(persistenceAdapter)
+    .withPersistenceAdapter(util.getPersistenceAdapter())
     .lambda();
