+++ 06/lambda/index.js
@@ -1,5 +1,5 @@
 const Alexa = require('ask-sdk-core');
-const persistence = require('./persistence');
+const util = require('./util');
 const interceptors = require('./interceptors');
 const logic = require('./logic');
 
@@ -341,6 +341,6 @@ exports.handler = Alexa.SkillBuilders.custom()
     .addResponseInterceptors(
         interceptors.LoggingResponseInterceptor,
         interceptors.SaveAttributesResponseInterceptor)
-    .withPersistenceAdapter(persistence.getPersistenceAdapter())
+    .withPersistenceAdapter(util.getPersistenceAdapter())
     .withApiClient(new Alexa.DefaultApiClient())
     .lambda();
